+++ src/Resources/contao/languages/en/tl_newsletter.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * DSB-Newsletter für Contao
+ *
+ * @author    Frank Hoppe
+ * @license   LGPL-3.0-or-later
+ * @copyright Frank Hoppe
+ */
+
+/**
+ * Fields
+ */
+$GLOBALS['TL_LANG']['tl_newsletter']['dsbItems'] = array('Use content elements', 'Allows you to build the newsletter from sortable sections with a headline, an image and a text. When you enable this, the insert tag {{newsletter::ID}} is added to the end of the HTML content unless one is already present. It marks the spot where the sections appear and can be moved freely.');
+
+/**
+ * Buttons
+ */
+$GLOBALS['TL_LANG']['tl_newsletter']['editContent'] = array('Edit content elements', 'Edit the content elements of newsletter %s');